+++ database/migrations/2023_03_05_084549_create_import_sales_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('import_sales', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('customer_id');
-            $table->unsignedBigInteger('amount');
-            $table->unsignedBigInteger('debt')->default(0);
+            $table->double('amount');
+            $table->double('debt')->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
